move mysql scanner to its own module
